+++ include/displaylib_LED_PICO/tm1637.hpp
@@ -18,11 +18,11 @@
 /*!
 	@brief Class for TM1637 Model 4
 */
-class TM1637plus_model4 : public SevenSegmentFont , public CommonData{
+class TM1637plus_modelX : public SevenSegmentFont , public CommonData{
 
 public:
 
-	TM1637plus_model4 (uint8_t clock, uint8_t data , int delay, int DisplaySize) ;
+	TM1637plus_modelX (uint8_t clock, uint8_t data , int delay, int DisplaySize) ;
 	void displayBegin(void);
 	void displayClose(void);
 	void displayClear(void);