+++ extra/doc/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 * Version 1.0.0 August 2022
-	* first release, tm1638
+	* first release, tm1638 only.
 
 * Version 1.9.0 October 2022
 	* Updating project so it's matches changes made to original Arduino port, V1.9.0.
@@ -13,3 +13,7 @@
 	* Changed name of project to displaylib_LED_PICO, it will now support all
 	LED segment displays that will be added
 	* Added support for ht16k33, tm1637 & max7219
+
+* Version 2.2.0 July 2025
+	* Corrected code so Max7219 cascade examples works, just acquired second display
+	* Added tm1637 example for version with decimal points in addition to semi-colon version